refactor(pms): route wms inventory reads through integration client
